refactor: reorganize request manager into package
