--- lightllm/common/req_manager.py
@@ -1,327 +1,0 @@
-import torch
-import collections
-from lightllm.common.linear_att_cache_manager.config_objs import LinearAttCacheConfig
-
-from lightllm.utils.log_utils import init_logger
-from .kv_cache_mem_manager import MemoryManager
-from typing import List, Optional, TYPE_CHECKING
-from lightllm.common.basemodel.triton_kernel.gen_sampling_params import token_id_counter
-from lightllm.common.basemodel.triton_kernel.gen_sampling_params import update_req_to_token_id_counter
-from lightllm.utils.envs_utils import get_env_start_args
-from lightllm.utils.config_utils import get_vocab_size
-from lightllm.server.router.model_infer.pin_mem_manager import g_pin_mem_manager
-from lightllm.common.linear_att_cache_manager.layer_cache import LayerCache
-from lightllm.common.linear_att_cache_manager.linear_att_buffer_manager import LinearAttCacheManager
-
-if TYPE_CHECKING:
-    from lightllm.server.router.model_infer.infer_batch import InferReq
-
-logger = init_logger(__name__)
-
-
-class _ReqNode:
-    def __init__(self, index):
-        self.index = index
-        self.next: "_ReqNode" = None
-
-
-class _ReqLinkedList:
-    def __init__(self, max_request_num):
-        self.nodes = [_ReqNode(i) for i in range(max_request_num)]
-        self.marks = [0 for _ in range(max_request_num)]
-        self.root_node = _ReqNode(-1)
-        for i in range(0, max_request_num - 1):
-            self.nodes[i].next = self.nodes[i + 1]
-        self.root_node.next = self.nodes[0]
-        self.can_alloc_size = max_request_num
-        return
-
-    def alloc(self):
-        if self.root_node.next is None:
-            logger.warning("alloc req index fail")
-            return None
-        get_node = self.root_node.next
-        self.root_node.next = self.root_node.next.next
-        assert self.marks[get_node.index] == 0
-        self.marks[get_node.index] = 1
-        self.can_alloc_size -= 1
-        return get_node.index
-
-    def free(self, index):
-        assert self.marks[index] == 1
-        node = self.nodes[index]
-        node.next = self.root_node.next
-        self.root_node.next = node
-        self.marks[index] = 0
-        self.can_alloc_size += 1
-        return
-
-    def is_all_free(self):
-        return self.can_alloc_size == len(self.marks)
-
-
-class ReqManager:
-    def __init__(self, max_request_num, max_sequence_length, mem_manager: MemoryManager):
-        # 这里对最大请求数量的管理在默认上多申请了一个，主要是 index 为 max_request_num 代表
-        # 的这个请求管理 id， 主要是为了兼容 DP 运行模式下，让各个 DP 能 padding 到 DP 中最大
-        # 的那个batch size 进行运行，所有 padding 的请求都会使用预留的这个请求管理 id 进行处理
-        # 这样让 DP 的实现更为简化一些。
-        self.req_list = _ReqLinkedList(max_request_num)
-        self.req_to_token_indexs = torch.zeros(
-            (max_request_num + 1, max_sequence_length), dtype=torch.int32, device="cuda"
-        )
-        self.mem_manager = mem_manager
-        self.req_sampling_params_manager = ReqSamplingParamsManager(max_request_num)
-        self.max_request_num = max_request_num
-        self.HOLD_REQUEST_ID = max_request_num
-
-    def alloc(self):
-        return self.req_list.alloc()
-
-    def free(self, free_req_indexes: List[int], free_token_index):
-        for req_index in free_req_indexes:
-            self.req_list.free(req_index)
-
-        if self.req_list.is_all_free():
-            logger.debug(f"freed all request size {self.req_list.can_alloc_size}")
-        self.mem_manager.free(free_token_index)
-
-    def free_req(self, free_req_index: int):
-        self.req_list.free(free_req_index)
-        if self.req_list.is_all_free():
-            logger.debug(f"freed all request size {self.req_list.can_alloc_size}")
-        return
-
-    def free_token(self, free_token_index):
-        self.mem_manager.free(free_token_index)
-        return
-
-    def free_all(self):
-        self.req_list = _ReqLinkedList(self.max_request_num)
-        return
-
-
-class ReqSamplingParamsManager:
-    """
-    ReqSamplingParamsManager 将输出采样参数中，确定比较固定的部分，纳入到 gpu buffer中进行管理，这样可以更快捷的
-    利用 triton kernel 进行处理，对于那些比较动态(部分处理模式下会动态的修改某些后处理参数)，或者存在特殊处理的后处理参数，
-    则保留从 InferSamplingParams 中进行动态读取和动态组batch， 具体使用可以参考
-    lightllm/server/router/model_infer/mode_backend/generic_post_process.py 文件中的使用方式。
-    """
-
-    def __init__(self, max_request_num):
-        # mode ["cpu_counter", "pin_mem_counter", "gpu_counter"]
-        self.penalty_counter_mode = get_env_start_args().penalty_counter_mode
-        self.vocab_size = get_vocab_size(get_env_start_args().model_dir)
-        self.mtp_verify_width = get_env_start_args().mtp_step + 1
-        self.req_to_presence_penalty = torch.zeros(max_request_num + 1, dtype=torch.float32, device="cuda")
-        self.req_to_frequency_penalty = torch.zeros(max_request_num + 1, dtype=torch.float32, device="cuda")
-        self.req_to_repetition_penalty = torch.zeros(max_request_num + 1, dtype=torch.float32, device="cuda")
-        self.req_to_next_token_ids = torch.zeros(
-            (max_request_num + 1, self.mtp_verify_width),
-            dtype=torch.int64,
-            device="cuda",
-        )
-        self.req_to_next_token_scores = (
-            torch.zeros_like(self.req_to_next_token_ids, dtype=torch.float32)
-            if get_env_start_args().mtp_dynamic_verify
-            else None
-        )
-
-        self.req_to_exponential_decay_length_penalty = torch.zeros(
-            max_request_num + 1, dtype=torch.float32, device="cuda"
-        )
-
-        if self.penalty_counter_mode == "gpu_counter":
-            self.req_to_out_token_id_counter = torch.zeros(
-                (max_request_num + 1, self.vocab_size), dtype=torch.int32, device="cuda"
-            )
-        elif self.penalty_counter_mode == "pin_mem_counter":
-            self.req_to_out_token_id_counter = torch.zeros(
-                (max_request_num + 1, self.vocab_size), dtype=torch.int32, device="cpu", pin_memory=True
-            )
-
-    def init_req_sampling_params(self, req: "InferReq"):
-        shm_param = req.sampling_param.shm_param
-        self.req_to_next_token_ids[req.req_idx][0:1].fill_(req.get_last_gen_token())
-        if self.req_to_next_token_scores is not None:
-            self.req_to_next_token_scores[req.req_idx].fill_(0.0)
-            self.req_to_next_token_scores[req.req_idx][0:1].fill_(1.0)
-        self.req_to_presence_penalty[req.req_idx].fill_(shm_param.presence_penalty)
-        self.req_to_frequency_penalty[req.req_idx].fill_(shm_param.frequency_penalty)
-        self.req_to_repetition_penalty[req.req_idx].fill_(shm_param.repetition_penalty)
-        exponential_decay_length_penalty = shm_param.exponential_decay_length_penalty.to_tuple()
-        self.req_to_exponential_decay_length_penalty[req.req_idx].fill_(exponential_decay_length_penalty[1])
-        # 提前标记当前请求是否需要统计输出token的计数，因为这个统计可能会导致一些特定场景下后处理效率的下降
-        # 所以提前标记不需要进行后处理统计的场景。
-        req.need_out_token_id_statistics = not (
-            shm_param.presence_penalty == 0.0
-            and shm_param.frequency_penalty == 0.0
-            and shm_param.repetition_penalty == 1.0
-        )
-
-        if self.penalty_counter_mode == "cpu_counter":
-            if req.sampling_param.shm_param.input_penalty and req.need_out_token_id_statistics:
-                req.out_token_id_count = collections.Counter(req.shm_req.get_prompt_ids())
-            else:
-                req.out_token_id_count = collections.defaultdict(int)
-        else:
-            self.req_to_out_token_id_counter[req.req_idx].fill_(0)
-            if req.sampling_param.shm_param.input_penalty and req.need_out_token_id_statistics:
-                prompt_ids = g_pin_mem_manager.gen_from_list(
-                    key="prompt_ids_for_penalty",
-                    data=req.shm_req.get_prompt_ids_numpy(),
-                    dtype=torch.int32,
-                ).cuda(non_blocking=True)
-                token_id_counter(
-                    prompt_ids=prompt_ids, out_token_id_counter=self.req_to_out_token_id_counter[req.req_idx]
-                )
-                torch.cuda.current_stream().synchronize()
-
-        return
-
-    def update_reqs_out_token_counter_gpu(
-        self, b_req_idx: torch.Tensor, next_token_ids: torch.Tensor, mask: torch.Tensor = None
-    ):
-        if self.penalty_counter_mode not in ["gpu_counter", "pin_mem_counter"]:
-            return
-
-        assert b_req_idx.is_cuda and next_token_ids.is_cuda and b_req_idx.shape[0] == next_token_ids.shape[0]
-
-        update_req_to_token_id_counter(
-            b_req_idx=b_req_idx,
-            next_token_ids=next_token_ids,
-            req_to_out_token_id_counter=self.req_to_out_token_id_counter,
-            mask=mask,
-        )
-        return
-
-    def update_reqs_token_counter(
-        self, req_objs: List["InferReq"], next_token_ids: List[int], accept_mark: Optional[List[List[bool]]] = None
-    ):
-        if self.penalty_counter_mode != "cpu_counter":
-            return
-
-        for req_obj, next_token_id in zip(req_objs, next_token_ids):
-            if req_obj.need_out_token_id_statistics and req_obj.cur_output_len > 0:
-                req_obj.out_token_id_count[next_token_id] += 1
-        return
-
-    def gen_cpu_out_token_counter_sampling_params(self, req_objs: List["InferReq"]):
-        assert self.penalty_counter_mode == "cpu_counter"
-
-        p_token_ids: List[int] = []
-        p_token_counts: List[int] = []
-        p_cumsum_seq_len: List[int] = [
-            0,
-        ]
-        cum_sum_len = 0
-        for i, req_obj in enumerate(req_objs):
-            id_to_count = req_obj.out_token_id_count
-            p_token_ids.extend(list(id_to_count.keys()))
-            p_token_counts.extend(list(id_to_count.values()))
-            cum_sum_len += len(id_to_count)
-            p_cumsum_seq_len.append(cum_sum_len)
-
-        p_token_ids_tensor = g_pin_mem_manager.gen_from_list(key="p_token_ids", data=p_token_ids, dtype=torch.int32)
-        p_token_counts_tensor = g_pin_mem_manager.gen_from_list(
-            key="p_token_counts", data=p_token_counts, dtype=torch.int32
-        )
-        p_cumsum_seq_len_tensor = g_pin_mem_manager.gen_from_list(
-            key="p_cumsum_seq_len", data=p_cumsum_seq_len, dtype=torch.int32
-        )
-
-        return (
-            p_token_ids_tensor.cuda(non_blocking=True),
-            p_token_counts_tensor.cuda(non_blocking=True),
-            p_cumsum_seq_len_tensor.cuda(non_blocking=True),
-        )
-
-
-class ReqManagerForMamba(ReqManager):
-    def __init__(self, max_request_num, max_sequence_length, mem_manager, linear_config: LinearAttCacheConfig):
-        super().__init__(max_request_num, max_sequence_length, mem_manager)
-        self.mtp_step = get_env_start_args().mtp_step
-        # 因为在mtp的推理中，需要标记每个请求对应的mtp index状态(conv state 和 ssm state)，在mtp对应序列中
-        # 的真实位置，所以需要需要一个标记来记录，不然算子无法找到真实的处理起点。
-        self.req_to_mtp_state_index = (
-            torch.zeros((max_request_num + 1,), dtype=torch.int32, device="cuda") if self.mtp_step > 0 else None
-        )
-        # 突然想到， 在linear att 开启mtp的模式中，现在的prefill linear att 算子默认是从0的位置读取信息进行操作
-        # 所以不能支持 prefill decode mixed 操作了，因为一个decode过的请求，重新用prefill 算子跑，会出现读错linear
-        # 状态位置的问题。导致bug, 在这里加个断言，以后可以支持上 TODO
-        if self.mtp_step > 0:
-            assert get_env_start_args().enable_prefill_decode_mixed is False
-
-        self.big_page_token_num = (
-            get_env_start_args().linear_att_page_block_num * get_env_start_args().linear_att_hash_page_size
-        )
-        self.linear_config = linear_config
-
-        self.req_to_conv_state = LayerCache(
-            size=(max_request_num + 1),
-            dtype=self.linear_config.conv_state_dtype,
-            shape=self.linear_config.get_mtp_conv_state_shape(mtp_step=self.mtp_step),
-            layer_num=self.linear_config.linear_layer_num,
-            device="cuda",
-        )
-        self.req_to_ssm_state = LayerCache(
-            size=(max_request_num + 1) * (self.mtp_step + 1),
-            dtype=self.linear_config.ssm_state_dtype,
-            shape=self.linear_config.get_ssm_state_shape(),
-            layer_num=self.linear_config.linear_layer_num,
-            device="cuda",
-        )
-        return
-
-    def init_linear_att_state(self, req: "InferReq"):
-        conv_index = req.req_idx
-        ssm_start = req.req_idx * (self.mtp_step + 1)
-        self.req_to_conv_state.buffer[:, conv_index, ...].fill_(0)
-        # #17: zero the FULL (mtp_step + 1)-row SSM block, not just canonical row +0, so a future
-        # first-step verify reading offset>0 after fresh init never hits a never-written row (NaN).
-        self.req_to_ssm_state.buffer[:, ssm_start : ssm_start + (self.mtp_step + 1), ...].fill_(0)
-        if self.req_to_mtp_state_index is not None:
-            self.req_to_mtp_state_index[req.req_idx] = 0
-        return
-
-    def get_mamba_cache(self, layer_idx_in_all: int):
-        assert (
-            0 <= layer_idx_in_all < self.linear_config.all_layer_num
-        ), f"invalid transformer layer index {layer_idx_in_all}"
-        layer_idx_in_linear = layer_idx_in_all - (layer_idx_in_all // self.linear_config.full_attention_interval)
-        conv_states = self.req_to_conv_state.buffer[layer_idx_in_linear]
-        ssm_states = self.req_to_ssm_state.buffer[layer_idx_in_linear]
-        return conv_states, ssm_states
-
-    def copy_big_page_buffer_to_linear_att_state(self, big_page_buffer_idx: int, req: "InferReq"):
-        from .linear_att_cache_manager import LinearAttCacheManager
-
-        big_page_buffers: LinearAttCacheManager = self.mem_manager.linear_att_big_page_buffers
-
-        conv_state, ssm_state = big_page_buffers.get_state_cache(buffer_idx=big_page_buffer_idx)
-        conv_dest = req.req_idx
-        ssm_dest = req.req_idx * (self.mtp_step + 1)
-        conv_cache_width = conv_state.shape[-1]
-        self.req_to_conv_state.buffer[:, conv_dest, ..., :conv_cache_width] = conv_state
-        self.req_to_ssm_state.buffer[:, ssm_dest, ...] = ssm_state
-        if self.req_to_mtp_state_index is not None:
-            self.req_to_mtp_state_index[req.req_idx] = 0
-        return
-
-    def copy_small_page_buffer_to_linear_att_state(
-        self, req: "InferReq", linear_att_small_page_buffers: LinearAttCacheManager
-    ):
-        conv_state, ssm_state = linear_att_small_page_buffers.get_state_cache(
-            buffer_idx=req.shared_kv_node.small_page_buffer_idx
-        )
-        conv_dest = req.req_idx
-        ssm_dest = req.req_idx * (self.mtp_step + 1)
-        conv_cache_width = conv_state.shape[-1]
-        # TODO 下面这个从 cpu cache 拷贝数据的 gpu的操作，是否是阻塞的操作。
-        # 同时，非连续对象的拷贝，可能存在效率问题。
-        self.req_to_conv_state.buffer[:, conv_dest, ..., :conv_cache_width] = conv_state
-        self.req_to_ssm_state.buffer[:, ssm_dest, ...] = ssm_state
-        if self.req_to_mtp_state_index is not None:
-            self.req_to_mtp_state_index[req.req_idx] = 0
-        return
